[DAEF-300] fixes storybook setup
