+++ app/Routes.js
@@ -1,6 +1,7 @@
 // @flow
 import React from 'react';
 import { Route, IndexRedirect } from 'react-router';
+import { ROUTES } from './routes-config';
 
 // PAGES
 import Wallet from './containers/wallet/Wallet';
@@ -21,34 +22,7 @@ import TermsOfUseSettingsPage from './containers/settings/categories/TermsOfUseS
 import TermsOfUsePage from './containers/profile/TermsOfUsePage';
 import SendLogsChoicePage from './containers/profile/SendLogsChoicePage';
 
-export const ROUTES = {
-  ROOT: '/',
-  STAKING: '/staking',
-  ADA_REDEMPTION: '/ada-redemption',
-  NO_WALLETS: '/no-wallets',
-  PROFILE: {
-    LANGUAGE_SELECTION: '/profile/language-selection',
-    TERMS_OF_USE: '/profile/terms-of-use',
-    SEND_LOGS: '/profile/send-logs-choice',
-  },
-  WALLETS: {
-    ROOT: '/wallets',
-    PAGE: '/wallets/:id/:page',
-    SUMMARY: '/wallets/:id/summary',
-    TRANSACTIONS: '/wallets/:id/transactions',
-    SEND: '/wallets/:id/send',
-    RECEIVE: '/wallets/:id/receive',
-    SETTINGS: '/wallets/:id/settings',
-  },
-  SETTINGS: {
-    ROOT: '/settings',
-    GENERAL: '/settings/general',
-    TERMS_OF_USE: '/settings/terms-of-use',
-    SUPPORT: '/settings/support',
-  },
-};
-
-export default (
+export const Routes = (
   <div>
     <Route path={ROUTES.ROOT} component={LoadingPage} />
     <Route path={ROUTES.PROFILE.LANGUAGE_SELECTION} component={LanguageSelectionPage} />
