+++ app/components/layout/TopBar.js
@@ -9,7 +9,7 @@ import menuIconOpened from '../../assets/images/menu-second-level-ic.svg';
 import menuIconClosed from '../../assets/images/menu-ic.svg';
 import styles from './TopBar.scss';
 import { matchRoute } from '../../lib/routing-helpers';
-import { ROUTES } from '../../Routes';
+import { ROUTES } from '../../routes-config';
 
 @observer
 export default class TopBar extends Component {
