+++ app/routes-config.js
@@ -1,0 +1,27 @@
+// @flow
+export const ROUTES = {
+  ROOT: '/',
+  STAKING: '/staking',
+  ADA_REDEMPTION: '/ada-redemption',
+  NO_WALLETS: '/no-wallets',
+  PROFILE: {
+    LANGUAGE_SELECTION: '/profile/language-selection',
+    TERMS_OF_USE: '/profile/terms-of-use',
+    SEND_LOGS: '/profile/send-logs-choice',
+  },
+  WALLETS: {
+    ROOT: '/wallets',
+    PAGE: '/wallets/:id/:page',
+    SUMMARY: '/wallets/:id/summary',
+    TRANSACTIONS: '/wallets/:id/transactions',
+    SEND: '/wallets/:id/send',
+    RECEIVE: '/wallets/:id/receive',
+    SETTINGS: '/wallets/:id/settings',
+  },
+  SETTINGS: {
+    ROOT: '/settings',
+    GENERAL: '/settings/general',
+    TERMS_OF_USE: '/settings/terms-of-use',
+    SUPPORT: '/settings/support',
+  },
+};
